Even further considerations
Account for once the Browse Patterns button/anchor on the home page is too big for the view and some patterns have their text cut off (about 300px). `min-width: 150px` needs to be wide enough to hold any (reasonable) word but should not interfere with the flowing layout. Meanwhile `5vw` keeps the Browse Patterns button/anchor at the same size as the headings.